+++ frontend/src/hooks/api/getters/useSignalQuery/useSignalQuery.ts
